chore(benches): restructure benchmarks into image and audio
